webassembly-stubber/requirements-test.txt
# https://playwright.dev/python/docs/intro
pytest-playwright

# note: also needs 
# playwright install 